Django_Server/requirements.txt
aioredis==1.3.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
Click==7.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
daphne==2.4.1
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.41.0
django-appconf==1.0.3
django-filter==2.2.0
django-google-auth==0.4.0
django-imagekit==4.0.2
django-rest-swagger==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.0
Flask==1.1.1
flex==6.14.1
google==2.0.3
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-oauth==1.0.1
hiredis==1.0.1
httplib2==0.15.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
inflection==0.3.1
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.3
jsonpointer==2.0
Markdown==3.1.1
MarkupSafe==1.1.1
msgpack==0.6.2
mysqlclient==1.4.6
oauth2client==4.1.3
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==19.2
pilkit==2.0
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-magic==0.4.15
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rfc3987==1.3.8
rsa==4.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
scrapy-googleauth==0.1.0
service-identity==18.1.0
simplejson==3.17.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.5
sqlparse==0.3.0
strict-rfc3339==0.7
Twisted==20.3.0
txaio==20.3.1
uritemplate==3.0.1
urllib3==1.25.7
validate-email==1.3
virtualenv==16.7.9
Werkzeug==0.16.0
zope.interface==5.0.1
